Renamed surrogate gradient. Added Identity
