+++ nir/ir/__init__.py
@@ -7,7 +7,7 @@ from .graph import Input, NIRGraph, Output
 from .linear import Affine, Linear, Scale
 from .neuron import IF, LI, LIF, CubaLIF, I
 from .pooling import AvgPool2d, SumPool2d
-from .surrogate_gradient import Threshold
+from .threshold import Threshold
 from .typing import NIRNode
 
 # NIRNodes that can be (de)serialized
